fix: Move the components around for better structure.
